convert to use CancelToken from libary
