+++ p2p/nat.py
@@ -9,7 +9,7 @@ from typing import (
 )
 from urllib.parse import urlparse
 
-from p2p.cancel_token import (
+from cancel_token import (
     CancelToken,
 )
 from p2p.exceptions import (
