+++ p2p/state.py
@@ -30,6 +30,8 @@ from eth_typing import (
     Hash32
 )
 
+from cancel_token import CancelToken
+
 from eth.constants import (
     BLANK_ROOT_HASH,
     EMPTY_SHA3,
@@ -40,7 +42,6 @@ from eth.rlp.accounts import Account
 from p2p import eth
 from p2p import protocol
 from p2p.chain import PeerRequestHandler
-from p2p.cancel_token import CancelToken
 from p2p.exceptions import OperationCancelled
 from p2p.peer import ETHPeer, PeerPool, PeerSubscriber
 from p2p.service import BaseService
