+++ trinity/rpc/ipc.py
@@ -10,10 +10,10 @@ from typing import (
 
 from cytoolz import curry
 
-from p2p.cancel_token import (
+from cancel_token import (
     CancelToken,
-    wait_with_token,
 )
+
 from p2p.exceptions import (
     OperationCancelled,
 )
@@ -61,17 +61,17 @@ async def connection_loop(execute_rpc: Callable[[Any], Any],
     while True:
         request_bytes = b''
         try:
-            request_bytes = await wait_with_token(reader.readuntil(b'}'), token=cancel_token)
+            request_bytes = await cancel_token.cancellable_wait(reader.readuntil(b'}'))
         except asyncio.LimitOverrunError as e:
             logger.info("Client request was too long. Erasing buffer and restarting...")
-            request_bytes = await wait_with_token(reader.read(e.consumed), token=cancel_token)
-            await wait_with_token(write_error(
+            request_bytes = await cancel_token.cancellable_wait(reader.read(e.consumed))
+            await cancel_token.cancellable_wait(write_error(
                 writer,
                 "reached limit: %d bytes, starting with '%s'" % (
                     e.consumed,
                     request_bytes[:20],
                 ),
-            ), token=cancel_token)
+            ))
             continue
 
         raw_request += request_bytes.decode()
@@ -79,9 +79,8 @@ async def connection_loop(execute_rpc: Callable[[Any], Any],
         bad_prefix, raw_request = strip_non_json_prefix(raw_request)
         if bad_prefix:
             logger.info("Client started request with non json data: %r", bad_prefix)
-            await wait_with_token(
+            await cancel_token.cancellable_wait(
                 write_error(writer, 'Cannot parse json: ' + bad_prefix),
-                token=cancel_token,
             )
 
         try:
@@ -96,9 +95,8 @@ async def connection_loop(execute_rpc: Callable[[Any], Any],
 
         if not request:
             logger.debug("Client sent empty request")
-            await wait_with_token(
+            await cancel_token.cancellable_wait(
                 write_error(writer, 'Invalid Request: empty'),
-                token=cancel_token,
             )
             continue
 
@@ -106,14 +104,13 @@ async def connection_loop(execute_rpc: Callable[[Any], Any],
             result = execute_rpc(request)
         except Exception as e:
             logger.exception("Unrecognized exception while executing RPC")
-            await wait_with_token(
+            await cancel_token.cancellable_wait(
                 write_error(writer, "unknown failure: " + str(e)),
-                token=cancel_token,
             )
         else:
             writer.write(result.encode())
 
-        await wait_with_token(writer.drain(), token=cancel_token)
+        await cancel_token.cancellable_wait(writer.drain())
 
 
 def strip_non_json_prefix(raw_request: str) -> Tuple[str, str]:
